+++ c/python/obj.cc
@@ -131,6 +131,7 @@ bool _obj::is_list_or_tuple() const noexcept { return is_list() || is_tuple(); }
 bool _obj::is_int()           const noexcept { return v && PyLong_Check(v) && !is_bool(); }
 bool _obj::is_float()         const noexcept { return v && PyFloat_Check(v); }
 bool _obj::is_string()        const noexcept { return v && PyUnicode_Check(v); }
+bool _obj::is_bytes()         const noexcept { return v && PyBytes_Check(v); }
 bool _obj::is_type()          const noexcept { return v && PyType_Check(v); }
 bool _obj::is_ltype()         const noexcept { return v && Py_TYPE(v) == py_ltype; }
 bool _obj::is_stype()         const noexcept { return v && Py_TYPE(v) == py_stype; }