+++ CHANGELOG.md
@@ -42,6 +42,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `datatable` now exposes C API, to allow other C/C++ libraries interact
   with datatable Frames directly (#1469).
 
+- The column selector in `DT[i, j]` can now be a list of booleans of length
+  `DT.ncols`. The True/False entries in this list will indicate whether to
+  select the corresponding column or not. This can be used to implement a
+  simple column filter, for example:
+    ```
+    del DT[:, (name.endswith("_tmp") for name in DT.names)]
+    ```
+
 
 #### Fixed